Update 1.1.1. Segment Tree.cpp

+++ Reference Codes/1. Data Structure/1.1. Segment Tree/1.1.1. Segment Tree.cpp
@@ -14,7 +14,7 @@ struct Seg {
 		for (int i = flag; i < flag + N; i++) cin >> t[i];
 		for (int i = flag - 1; i >= 1; i--) t[i] = t[i << 1] + t[i << 1 | 1];
 	}
-	void modify(int p, int value) {  // set value at position p
+	void modify(int p, ll value) {  // set value at position p
 		for (t[p += flag - 1] = value; p > 1; p >>= 1) t[p >> 1] = t[p] + t[p ^ 1];
 	}
 	ll query(int l, int r) {
